fix unittest fix cmake test=develop
