+++ python/paddle/fluid/tests/unittests/test_eager_deletion_gru_net.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import unittest
-from test_eager_deletion_lstm_net import TestBase
+from test_eager_deletion_dynamic_rnn_base import TestBase
 import paddle.fluid as fluid
